Xcode内のファイル構造の整理
