+++ iOSEngineerCodeCheck/App/GitHubSearcherState.swift
