+++ iOSEngineerCodeCheck/Detail/ImageCache.swift
@@ -42,18 +42,3 @@ final class ImageCache: ImageCacheForPresenter {
         }
     }
 }
-
-@MainActor
-final class ImageCacheFactory {
-    private let isTest: Bool
-
-    init(isTest: Bool) {
-        self.isTest = isTest
-    }
-
-    func makeImageCache() -> ImageCache {
-        let downloader: DownloaderForImageCache =
-            isTest ? ImageDownloaderUITestMock() : ImageDownloader()
-        return ImageCache(downloader: downloader)
-    }
-}
