+++ iOSEngineerCodeCheck/Detail/ImageCacheState.swift
