+++ iOSEngineerCodeCheck/Global/AppDelegate.swift
